backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
PyJWT==2.8.1
werkzeug==3.0.1
psycopg2-binary==2.9.9
pdf2image==1.16.3
pytesseract==0.3.10
python-multipart==0.0.6
requests==2.31.0
openai==1.3.0
anthropic==0.7.0
email-validator==2.1.0
gunicorn==21.2.0
